refactor: add useSemesterDirectoriesStore hook
