+++ src/services/contents/note/NoteEditPage.tsx
@@ -1,4 +1,4 @@
-﻿import { useMemo, useState } from "react";
+﻿import { useState } from "react";
 import { useParams, useNavigate, useLocation } from "react-router-dom";
 import MDEditor from "@uiw/react-md-editor";
 import { getCodeString } from "rehype-rewrite";
@@ -7,13 +7,11 @@ import katex from "katex";
 import TextSection from "@/commons/section/TextSection";
 import Button from "@/commons/inputs/Button";
 import TagItem from "@/commons/tag/TagItem";
-
 import { IconArrowLeft } from "@/assets";
 import type { TagData } from "@/types/tag";
 import { useTagStore } from "@/store/useTagStore";
-import { useDirectoryStore } from "@/store/useDirectoryStore";
 import { updateNoteContent, updateNoteTag, updateNoteTitle } from "@/api/Note";
-import { useSemesterStore } from "@/store/useSemesterStore";
+import { useSemesterDirectoriesStore } from "@/store/useSemesterDirectoriesStore";
 
 function NoteEditPage() {
   const navigate = useNavigate();
@@ -23,21 +21,7 @@ function NoteEditPage() {
   const directoryId = Number(slug?.split("-").pop());
   const noteId = Number(id);
   const allTags = useTagStore((state) => state.tags);
-
-  const { selectedSemesterKey } = useSemesterStore();
-
-  const { year, term } = useMemo(() => {
-    if (!selectedSemesterKey) return { year: 0, term: "" };
-    const [y, t] = selectedSemesterKey.split("-");
-    return { year: Number(y), term: t };
-  }, [selectedSemesterKey]);
-
-  const directories = useDirectoryStore((state) =>
-    state.getDirectoriesBySemester(year, term)
-  );
-
-  const studyDirectory = directories.find((dir) => dir.name === "학업");
-  const studyDirectoryId = studyDirectory?.id;
+  const { studyDirectoryId } = useSemesterDirectoriesStore();
 
   const originalTitle = location.state?.title || "";
   const originalContent = location.state?.markdownContent || "";
